wrap slide up/down functions in setTimeout
there's an issue specific to Firefox where hitting spacebar on a button was opening and closing the accordion panel immedietely.

by wrapping each of the checks with a short timeout function, this stopped happening.

this is not a good fix.  but one that at least fixes the problem for now...