updates. see desc.
remove js- classes and just use one set of classes for both stlying and js hooks.

fix spacebar bug with firefox where on buttons it was triggering the reveal/hide function twice.  this seemed to be an issue only on firefox, but the fix works on all browsers, so it should be a-ok going forward.